update category and unit tests

+++ src/main/kotlin/dev/janssenbatista/tqi_Kotlin_backedn_developer_2023/services/CategoryService.kt
@@ -3,6 +3,7 @@ package dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.services
 import dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.dtos.CategoryRequestDto
 import dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.exceptions.CategoryAlreadyExistsException
 import dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.exceptions.CategoryNotFoundException
+import dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.exceptions.ConstraintViolationException
 import dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.models.Category
 import dev.janssenbatista.tqi_Kotlin_backedn_developer_2023.repositories.CategoryRepository
 import org.springframework.data.domain.Page
@@ -51,7 +52,9 @@ class CategoryService(private val categoryRepository: CategoryRepository) {
         val category = categoryRepository.findById(categoryId).orElseThrow {
             throw CategoryNotFoundException("category not found")
         }
-
+        if (category.products.isNotEmpty()) {
+            throw ConstraintViolationException("cannot delete this category because are products associated")
+        }
         categoryRepository.delete(category)
     }
